Move labels-related content tests to a separate source file
